YKI(Frontend): move CustomerInformation to an own file
